refactor(v1/proxmox): extract shared host/PVE helpers into _helpers.py
